requirements.txt
torch>=1.12.0
torchxrayvision>=0.0.45
torchvision>=0.13.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
statsmodels>=0.14.0
pydicom>=2.4.0
tqdm>=4.65.0
pillow>=9.0.0
